prefactor: extract out tasks and helper functions
